Move torch.init to torch.nn.init to match PyTorch
